requirements.txt
fastapi==0.115.12
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.8.5
redis==5.0.1
sqlalchemy==2.0.40
alembic==1.15.2
psycopg2-binary==2.9.10
pydantic==2.11
httpx==0.27.0
pytest==8.3.5
asyncpg==0.30.0
python-multipart==0.0.20
